Rewrite Epub parsing with the new XmlParser implementation
Elements and attributes are every time picked from the right namespace.
Epub 3.x Property Data Type default vocabularies and prefix mechanism are carefully implemented.

As data are not organized in the same way in an Epub Package Document and in a Readium Publication,
a specific intermediate model is used to parse completely an Epub before converting it into a Publication.
This allows clearer code and should make further feature additions easier.