Assume UTC when parsing ISO 8601 dates (#144)
Assume that a date without a time zone component is in UTC. This ensures that apps see exactly the same dates (e.g. published) no matter where they are located.

For the same reason, the output Date will be in UTC. Apps should convert it to the local time zone for display purposes, or keep it as UTC for storage.