Rename `onLoadExternalResource` to `loadExternalResource`
The `onLoadExternalResource` callback in `WebViewServer` and its usage in `EpubNavigatorViewModel` has been renamed to `loadExternalResource`. This change provides a more descriptive name for the callback's function.

Additionally, a default empty implementation for `onExternalLinkActivated` in `HyperlinkNavigator.Listener` has been added. This makes implementing the listener more convenient as this method is often not needed.